refactor(PriceList): поменял нейминг компонентов
